feat: add indexing after adding

+++ src/repository/youtube-video-api.repository.ts
@@ -1,9 +1,10 @@
 import { connect } from "../config/db.config";
-import { Videos, VideosInterface } from "../model/youtube-video-api.model";
+import { videos, VideosInterface } from "../model/youtube-video-api.model";
 
 export class YoutubeVideoAPIRepository {
   private db: any = {};
   private youtubeVideoAPIRepository: any;
+  private sequelize: any;
 
   constructor() {
     this.db = connect();
@@ -11,13 +12,16 @@ export class YoutubeVideoAPIRepository {
     this.db.sequelize.sync({ force: false }).then(() => {
       console.log("Drop and re-sync db.");
     });
-    this.youtubeVideoAPIRepository = this.db.sequelize.getRepository(Videos);
+    this.youtubeVideoAPIRepository = this.db.sequelize.getRepository(videos);
+    this.sequelize = this.db.sequelize;
   }
 
   async getVideos() {
     try {
-      const videos = await this.youtubeVideoAPIRepository.findAll();
-      console.log("videos:::", videos);
+      const videos = await this.youtubeVideoAPIRepository.findAll({
+        attributes: { exclude: ["search_doc_weights"] },
+        order: [["published_at", "DESC"]],
+      });
       return videos;
     } catch (err) {
       console.log(err);
@@ -25,14 +29,21 @@ export class YoutubeVideoAPIRepository {
     }
   }
 
-  async saveVideos(toBeSaved: VideosInterface[]) {
+  async saveVideos(toBeSaved: VideosInterface[], num: number) {
     try {
-      await this.youtubeVideoAPIRepository.destroy({
-        where: {},
-        truncate: true,
-      });
+      if (num == 0) {
+        await this.youtubeVideoAPIRepository.destroy({
+          where: {},
+          truncate: true,
+        });
+      }
       const videos = await this.youtubeVideoAPIRepository.bulkCreate(toBeSaved);
-      console.log("videos:::", videos);
+      this.sequelize.query(
+        "update videos " +
+          "set search_doc_weights = setweight(to_tsvector(name), 'A') || setweight(to_tsvector(coalesce(description, '')), 'B'); " +
+          "drop index if exists search__weights_idx; " +
+          "create index search__weights_idx on videos using GIN(search_doc_weights);"
+      );
       return videos;
     } catch (err) {
       console.log(err);
